remove lerps

+++ Scripts/Currencies/Currency+Json.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using Dirichlet.Numerics;
 using UnityEngine;
 
 namespace Currencies
@@ -29,7 +30,7 @@ namespace Currencies
                 force = Prefs.GetInt(LegacyKey);
                 Prefs.DeleteKey(LegacyKey);
             }
-            if (force >= 0) _data.Total = force;
+            if (force >= 0) _data.Total = (ulong) force;
             _dirty = fresh || force >= 0;
             UpdateTracking();
         }
@@ -39,14 +40,14 @@ namespace Currencies
 
         private void SetInfinite(DateTime? until) => _data.InfiniteUntil = until?.ToBinary() ?? 0L;
 
-        private void SetTotal(long total) => _data.Total = total;
-        private long GetTotal() => _data.Total;
+        private void SetTotal(ulong total) => _data.Total = total;
+        private ulong GetTotal() => _data.Total;
 
-        private void SetStaged(long staged) => _data.Staged = staged;
-        private long GetStaged() => _data.Staged;
+        private void SetStaged(ulong staged) => _data.Staged = staged;
+        private ulong GetStaged() => _data.Staged;
 
-        private void SetLifetime(long lifetime) => _data.Lifetime = lifetime;
-        private long GetLifetime() => _data.Lifetime;
+        private void SetLifetime(UInt128 lifetime) => _data.Lifetime = lifetime;
+        private UInt128 GetLifetime() => _data.Lifetime;
 
         private void SetMultipliers(List<CurrencyModifier.Multiplier> mults) => _data.Multipliers = mults;
         private List<CurrencyModifier.Multiplier> GetMultipliers() => _data.Multipliers;
